Refactor: separate control buttons from header layout, export single ClerkTranslatorDetailsFields component
