+++ src/main/reactjs/src/interfaces/clerkTranslatorDetailsField.ts
@@ -9,7 +9,5 @@ import { CustomTextFieldProps } from 'interfaces/customTextField';
 export type ClerkTranslatorDetailsFieldProps = {
   translator?: ClerkTranslator;
   field: keyof ClerkTranslatorBasicInformation;
-  onFieldChange: (
-    e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-  ) => void;
+  onChange: (e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => void;
 } & CustomTextFieldProps;
